Revert "Dev"

+++ src/main/java/me/hapyl/fight/game/maps/NonPlayableGameMap.java
@@ -4,11 +4,11 @@ import org.bukkit.Material;
 
 public class NonPlayableGameMap extends GameMap {
 
-    public NonPlayableGameMap(String name, String description, int x, int y, int z) {
+    public NonPlayableGameMap(String name, String description, double x, double y, double z) {
         this(name, description, x, y, z, 0.0f, 0.0f);
     }
 
-    public NonPlayableGameMap(String name, String description, int x, int y, int z, float yaw, float pitch) {
+    public NonPlayableGameMap(String name, String description, double x, double y, double z, float yaw, float pitch) {
         super(name);
 
         setPlayable(false);
